Updated Pagination's input and output type to accept different types of input
Since we create a copy of the items passed to Pagination anyway, we might as well allow users to pass in any type that implements IEnumerable<T>.
This will make the Pagination type more versertile for users, who might not have a List<T> as the source for the data they are sending to the clients.
The Items property is changed to a IReadOnlyCollection since the receiver will not need to add any more items to the collection.