Simplify GetDotnetExecutable by converting check from Directory.GetFiles to File.Exists
When this method was first written, we didn't use the RuntimeInformation.IsOSPlatform to check if we were on a Windows OS.
We instead used the searchPattern "dotnet*" to match both "dotnet" and "dotnet.exe" and therefor it made sense to have a switch case where we validated how many files we matched.
Now that the filename is not a wildcard, we simply need to check if the file exists.